+++ HISTORY.rst
@@ -29,6 +29,7 @@ latest (unreleased)
   performance issue.
 
 * ALCYN-1972: Create filter for articles in BO on purchase
+* ALCYN-1965: Add the storage temperature when exporting a product to the ESB.
 
 **Bugfixes**